+++ sdk/CHANGELOG.md
@@ -72,6 +72,7 @@
 - Redesign `Client::send_msg` and `Relay::send_msg` APIs (https://github.com/rust-nostr/nostr/pull/1240)
 - Redesign `Client::notifications` and `Relay::notifications` APIs (https://github.com/rust-nostr/nostr/pull/1245)
 - Bump MSRV to 1.85.0 (https://github.com/rust-nostr/nostr/pull/1267)
+- Terminate event stream on drop
 
 ### Added